pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
